Consolidate user-role repository into rbac module
