+++ src/data/repositories/index.ts
@@ -5,4 +5,3 @@ export * from './refresh-token'
 export * from './team'
 export * from './token'
 export * from './user'
-export * from './user-role'
