+++ src/data/repositories/rbac/index.ts
@@ -1,10 +1,13 @@
-import { setUserPermissions } from './mutations'
-import { findRolePermissions, findUserPermissions } from './queries'
+import { assignRole, setUserPermissions } from './mutations'
+import { findInviters, findRole, findRolePermissions, findUserPermissions } from './queries'
 
 export * from './queries'
 export * from './types'
 
 export const rbacRepo = {
+  assignRole,
+  findInviters,
+  findRole,
   findRolePermissions,
   findUserPermissions,
   setUserPermissions,
