+++ src/data/repositories/rbac/queries.ts
@@ -1,11 +1,12 @@
-import { and, eq, isNotNull, isNull, or } from 'drizzle-orm'
+import { and, eq, inArray, isNotNull, isNull, or } from 'drizzle-orm'
 
 import type { Action, Resource, Role } from '@/types'
 
 import type { Database } from '../../clients'
+import type { Inviter, UserRoleRecord } from './types'
 
 import { db } from '../../clients'
-import { permissionsTable, rolePermissionsTable, userPermissionsTable } from '../../schema'
+import { permissionsTable, rolePermissionsTable, userPermissionsTable, userRolesTable, usersTable } from '../../schema'
 
 export interface ActivePermissionRow {
   action: Action
@@ -70,4 +71,45 @@ export const findUserPermissions = async (
         ),
       ),
     )
+}
+
+export const findRole = async (
+  userId: string,
+  tx?: Database,
+): Promise<UserRoleRecord | undefined> => {
+  const executor = tx || db
+
+  const [found] = await executor
+    .select()
+    .from(userRolesTable)
+    .where(eq(userRolesTable.userId, userId))
+
+  return found
+}
+
+export const findInviters = async (
+  userIds: string[],
+  tx?: Database,
+): Promise<Map<string, Inviter>> => {
+  if (userIds.length === 0) {
+    return new Map()
+  }
+
+  const executor = tx || db
+  const rows = await executor
+    .select({
+      userId: userRolesTable.userId,
+      inviterId: usersTable.id,
+      firstName: usersTable.firstName,
+      lastName: usersTable.lastName,
+    })
+    .from(userRolesTable)
+    .innerJoin(usersTable, eq(userRolesTable.invitedBy, usersTable.id))
+    .where(inArray(userRolesTable.userId, userIds))
+
+  return new Map(rows.map(r => [r.userId, {
+    id: r.inviterId,
+    firstName: r.firstName,
+    lastName: r.lastName,
+  }]))
 }
