--- src/data/repositories/user-role/index.ts
@@ -1,11 +1,0 @@
-import { assign, remove } from './mutations'
-import { findByUserId, findInviters } from './queries'
-
-export * from './types'
-
-export const userRoleRepo = {
-  assign,
-  findByUserId,
-  findInviters,
-  remove,
-}
